separate out the base class
