+++ tests/src/Unit/OgAccessEntityTest.php
@@ -18,7 +18,7 @@ use Drupal\og\OgAccess;
  * @group og
  * @coversDefaultClass \Drupal\og\OgAccess
  */
-class OgAccessEntityTest extends OgAccessTest  {
+class OgAccessEntityTest extends OgAccessTestBase  {
 
   protected $entity;
 
@@ -69,7 +69,18 @@ class OgAccessEntityTest extends OgAccessTest  {
    * @coversDefaultmethod ::userAccessEntity
    * @dataProvider operationProvider
    */
-  public function xtestUserAccessEntityNew($operation) {
+  public function testDefaultForbidden($operation) {
+    $group_entity = $this->groupEntity();
+    $group_entity->isNew()->willReturn(FALSE);
+    $user_access = OgAccess::userAccessEntity($operation, $this->entity->reveal(), $this->user->reveal());
+    $this->assertTrue($user_access->isForbidden());
+  }
+
+  /**
+   * @coversDefaultmethod ::userAccessEntity
+   * @dataProvider operationProvider
+   */
+  public function testEntityNew($operation) {
     $group_entity = $this->groupEntity();
     $group_entity->isNew()->willReturn(TRUE);
     $user_access = OgAccess::userAccessEntity($operation, $group_entity->reveal(), $this->user->reveal());
@@ -80,7 +91,7 @@ class OgAccessEntityTest extends OgAccessTest  {
    * @coversDefaultmethod ::userAccessEntity
    * @dataProvider operationProvider
    */
-  public function testUserAccessGetEntityGroups($operation) {
+  public function testGetEntityGroups($operation) {
     $this->user->hasPermission(OgAccess::ADMINISTER_GROUP_PERMISSION)->willReturn(TRUE);
     $user_entity_access = OgAccess::userAccessEntity($operation, $this->entity->reveal(), $this->user->reveal());
     $this->assertTrue($user_entity_access->isAllowed());
